examples/hello-smolix: Add wip "hello world" demo
This is not the end of the road. I'm cheating a bit by adding the full
minimal-initramfs content in there, rather than modularizing this a
bit... But really *it's enough for this demo*.

I'm thinking it makes no sense to force stage-1 to do busybox things and
similar. Really it should be an *additional module*, which is enabled by
default, that builds such a system. This way overriding it for something
else entirely is trivial.