wip/u-boot: copy boot.scr to recovery.scr
This is a workaround for the lack of proper semantics for users of the
module.

Either the filesystem content needs to be editable in some form by
consumer modules, and/or we allow configuring the name of the script
output used in here.

The workaround makes the `recovery.scr` file usable on a PineNote
runinng the WIP Tow-Boot build with the "weird" recovery partition
concept.