image-builder: Add support for squashfs
But squashfs is not really a filesystem! Really? Well, it's in the name,
but otherwise I think it looks more like an archive (tar, cpio) than a
filesystem, since it's kinda immutable. There are also other filesystem
semantics that don't apply (filesystem label, filesystem uuid), but
otherwise it's filesystemy enough that we consider it worthwhile to
build here.

Aaaand, Linux can and does like to boot from one. Especially useful when
building an immutable image that is e.g. burned to SPI. With such a
rootfs you can make robust and resilient systems.