wip/uefi: Add support for bundling a dtb file
This picks the first file found in the dtb list. There's not much that
can be done for platforms where more than one dtb file can represent the
hardware, other than preferring *not* to do wholesale dtb replacement.

It should be preferred to rely on the platform firmware provided FDT. If
this is not sufficient, overlaying at runtime [to be implemented] should be
preferred.